capsicum: make integer types consistent
For Capsicum sub-rights use the same integer types as for
the system calls they constrain, specifically:
 - ioctl(2) syscall has unsigned int command argument
 - fcntl(2) syscall has unsigned int command argument.

So use int types rather than long types.

This change introduces a variance from the FreeBSD 10.x kernel
API, which uses u_long / uint32_t for ioctl/fcntl subrights.

(Those types are roughly consistent with ioctl(2) and fcntl(2)
syscalls on that platform, which take u_long / int command arguments
respectively).

+++ include/linux/capsicum.h
@@ -8,9 +8,9 @@ struct file;
 /* Complete rights structure (primary and subrights). */
 struct capsicum_rights {
 	struct cap_rights primary;
-	unsigned long fcntls;  /* Only valid if CAP_FCNTL set in primary. */
-	long nioctls;  /* Only valid if CAP_IOCTL set in primary; -1 => all ioctls allowed */
-	unsigned long *ioctls;
+	unsigned int fcntls;  /* Only valid if CAP_FCNTL set in primary. */
+	int nioctls;  /* Only valid if CAP_IOCTL set in primary; -1 => all ioctls allowed */
+	unsigned int *ioctls;
 };
